Project_delivery_tracking/requirements.txt
black
pytest
psycopg2-binary
astronomer-cosmos>=1.0.3
protobuf==3.20.0
